v3

+++ src/main/java/es/com/kete1987/sportmonks/library/v3/model/match/StatisticsDataValue.java
@@ -1,12 +1,12 @@
 package es.com.kete1987.sportmonks.library.v3.model.match;
 
 public class StatisticsDataValue {
-    private Long value;
+    private Double value;
 
     public StatisticsDataValue() {
     }
 
     public Long getValue() {
-        return value;
+        return value.longValue();
     }
 }
